Refactor(navigation): Move navigation functions to library for use in other test subsets (RHOSR/RHOAD/...)
